fix(config): fix cyclic imports
